+++ nodejs/app.js
@@ -9,6 +9,8 @@ const dotenv = require('dotenv');
 dotenv.config();
 const pageRouter = require('./routes/page');
 const { sequelize } = require('./models');
+const { readAndSaveExcelData } = require('./models/excel');
+const db = require('./models/index');
 
 const app = express();
 app.set('port', process.env.PORT || 8007);
@@ -17,13 +19,14 @@ nunjucks.configure('views', {
     express: app,
     watch: true,
 });
-sequelize.sync({ force: false })
-    .then(() => {
-        console.log('데이터베이스 연결 성공');
-    })
-    .catch((err) => {
-        console.error(err);
-    });
+db.sequelize.sync().then(async () => {
+    console.log('데이터베이스 연결 성공');
+    await readAndSaveExcelData(); // 엑셀 데이터를 데이터베이스에 저장
+}).catch((err) => {
+    console.error(err);
+    console.log('데이터베이스 연결 에러');
+    process.exit();
+});
 
 app.use(morgan('dev'));
 app.use(express.static(path.join(__dirname, 'public')));
