+++ nodejs/models/property.js
@@ -1,36 +1,36 @@
-const Sequelize = require('sequelize');
+const { Model, DataTypes } = require('sequelize');
 
-class Property extends Sequelize.Model {
+class Property extends Model {
     static initiate(sequelize) {
-        Property.init({
+        this.init({
             id: {
-                type: Sequelize.BIGINT,
+                type: DataTypes.BIGINT,
                 autoIncrement: true,
                 primaryKey: true
             },
             postalCode: {
-                type: Sequelize.STRING(50),
+                type: DataTypes.STRING(50),
                 allowNull: false,
             },
             roadAddress: {
-                type: Sequelize.STRING(100),
+                type: DataTypes.STRING(100),
                 allowNull: true,
             },
             jibunAddress: {
-                type: Sequelize.STRING(100),
+                type: DataTypes.STRING(100),
                 allowNull: true,
             }
         }, {
-                sequelize,
-                timestamps: true,
-                underscored: false,
-                modelName: 'Property',
-                tableName: 'property',
-                paranoid: true,
-                charset: 'utf8',
-                collate: 'utf8_general_ci',
+            sequelize,
+            modelName: 'Property',
+            tableName: 'properties',
+            timestamps: true,
+            underscored: false,
+            paranoid: true,
+            charset: 'utf8',
+            collate: 'utf8_general_ci',
         });
     }
-};
+}
 
 module.exports = Property;
